Feature: We now compute profile similarities by position which allows us to generate scatterplots of one method against another with one point for each ungapped, canonical position in the MSAs. These plots are now generated (see cross_analyze_methods_and_benchmarks_by_position) and colored by the sequence entropy of that position in the natural sequences where the three entropy ranges (low, medium, high) are defined s.t. they all have roughly the same number of points.
Robustness:
We now allow there to be more than expectn files. This is a response to debug.py which queues up additional jobs in case of failure so we can end up with a surplus of structures for some cases. The first expectn files are taken from a sorted list so results should be deterministic unless the set of files changes.

Bug fixes:
Fixed a bug (the 'All' published methods inclusion was never implemented).
The setup script now calls fixbb methods "Fixed" rather than "fixbb". This was preventing their analysis in the method vs. method plots where we expected them to be named "Fixed".

Still unimplemented:
Per-position sequence entropy boxplots